backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
aiofiles>=23.2.1
Pillow>=10.1.0
pytesseract>=0.3.10
PyMuPDF>=1.23.8
pydub>=0.25.1
speechrecognition>=3.10.0
